+++ TodoMate/Data/Service/TodoOrderService.swift
@@ -1,32 +1,30 @@
 //
-//  UserInfoService.swift
+//  TodoOrderService.swift
 //  TodoMate
 //
-//  Created by hs on 1/28/25.
+//  Created by hs on 3/5/25.
 //
 
 import Foundation
 
 class TodoOrderService: TodoOrderServiceType {
-    private let userDefaults = UserDefaults.standard
-    private let todoOrderDateKey = Const.TodoOrderDateKey
-    private let todoOrderKey = Const.TodoOrderKey
+    private let todoRepository: TodoOrderRepositoryType
+    private let calendar = Calendar.current
     
-    func saveDate(_ date: Date) {
-        userDefaults.set(date, forKey: todoOrderDateKey)
+    init(todoRepository: TodoOrderRepositoryType = TodoOrderRepository()) {
+        self.todoRepository = todoRepository
     }
     
-    func loadDate() -> Date? {
-        userDefaults.object(forKey: todoOrderDateKey) as? Date
+    func saveOrder(_ order: [String], for date: Date) {
+        todoRepository.saveOrder(order)
+        todoRepository.saveDate(date)
     }
     
-    func saveOrder(_ order: [String]) {
-        userDefaults.set(order, forKey: todoOrderKey)
-    }
-    
-    func loadOrder() -> [String] {
-        userDefaults.array(forKey: todoOrderKey) as? [String] ?? []
+    func loadOrder(for date: Date) -> [String]? {
+        guard let lastSavedDate = todoRepository.loadDate(),
+              calendar.isDate(date, inSameDayAs: lastSavedDate) else {
+            return nil
+        }
+        return todoRepository.loadOrder()
     }
 }
-
-
